# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API FORMATIVO CESAR",
    "description": "API REST para gestión de una distribuidora de medicamentos",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000",
      "description": "Servidor Local"
    }
  ],
  "paths": {
...
    "/reportes/{id}": {
      "get": {
        "tags": ["REPORTES"],
        "summary": "Obtener datos por id",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Datos del reporte",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Reporte" } } }
          },
          "404": { "description": "Reporte no encontrado" }
        }
      },
      "put": {
        "tags": ["REPORTES"],
        "summary": "Actualizar un reporte",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Reporte" } } }
        },
        "responses": {
          "200": { "description": "Reporte actualizado correctamente" },
          "404": { "description": "Reporte no encontrado" }
        }
      },
      "delete": {
        "tags": ["REPORTES"],
        "summary": "Eliminar un reporte",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Reporte eliminado correctamente" },
          "404": { "description": "Reporte no encontrado" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Reporte": {
        "type": "object",
        "properties": {
          "rep_id": { "type": "string", "description": "ID del reporte" },
          "rep_tipo": { "type": "string", "description": "Tipo del reporte" },
          "rep_fecha": { "type": "string", "format": "date-time", "description": "Fecha del reporte" },
          "rep_parametros": { "type": "string", "description": "Parámetros del reporte" },
          "rep_usu_id_fk": { "type": "string", "description": "ID del usuario que generó el reporte" },
          "rep_resultados": { "type": "string", "description": "Resultados del reporte" }
        }
      },
...
    }
  }
}